Check the comments

+++ src/msg_service/RedisMsgService.h
@@ -11,6 +11,9 @@
 #include "AbstractMsgService.h"
 #include <hiredis/hiredis.h>
 
+/*
+ * A message service work with Redis.
+ * */
 class RedisMsgService: public AbstractMsgService {
 public:
     const std::string CLIENT = "client";
@@ -18,6 +21,10 @@ public:
 
     RedisMsgService(std::string redisServerIp, int redisServerPort, MsgManager* msgMgr);
     virtual ~RedisMsgService();
+
+    /*
+     * Start service, it will connect to the redis server and begin all threads.
+     * */
     void startService();
 
 private:
@@ -25,8 +32,8 @@ private:
     std::string redisServerIp;
     int redisServerPort;
 
-    static redisReply* rpushTo(redisContext* conn, const char* key, int keyLen, const char* value, int valueLen);
-    static redisReply* lpopFrom(redisContext* conn, const char* key, int keyLen);
+    static redisReply* rpushTo(redisContext* conn, const char* key, size_t keyLen, const char* value, size_t valueLen);
+    static redisReply* lpopFrom(redisContext* conn, const char* key, size_t keyLen);
 
     static void* recvThreadRunner(void *service_ptr);
     static void* sendThreadRunner(void *service_ptr);
